# api/api.yaml
swagger: "2.0"
info:
  description: "github link: https://github.com/TomasBerezny/mtaaProjekt"
  version: "1.0.0"
  title: "Kde si kamos API"
  servers:
    - url: "http://localhost:8000"
tags:
- name: "User"
  description: "Access users"
- name: "Activity"
  description: "Access activities"
- name: "Group"
  description: "Access group"
- name: "Categories"
  description: "Access categories"
- name: "Auth"
  description: "Access authentication"
- name: "Invite"
  description: "Access invites"
- name: "UsersActivities"
  description: "Access usersActivities"
- name: "UsersGroups"
  description: "Access usersGroups"
paths:
  /users/{userId}:
    get:
      tags:
        - "User"
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Succesfully returns one user with provided id"
          schema:
            type: object
            properties:
              _id:
                type: string
              username:
                type: string
              password:
                type: string
              first_name:
                type: string
              last_name:
                type: string
              profile_photo_path:
                type: string
              bio:
                type: string
              phone_number:
                type: string
              __v:
                type: integer
        "404":
          description: Returns error
          schema:
            type: object
            properties:
              message:
                type: string
    patch:
      tags:
        - "User"
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
        - in: body
          name: user
          description: New user attributes.
          schema:
            type: object
            properties:
              username:
                type: string
              password:
                type: string
              first_name:
                type: string
              last_name:
                type: string
              bio:
                type: string
              phone_number:
                type: string
      responses:
        "200":
          description: "Succesfully returns updated user"
          schema:
            type: object
            properties:
              _id:
                type: string
              username:
                type: string
              password:
                type: string
              first_name:
                type: string
              last_name:
                type: string
              profile_photo_path:
                type: string
              bio:
                type: string
              phone_number:
                type: string
              __v:
                type: integer
        "400":
          description: Returns error
          schema:
            type: object
            properties:
              message:
                type: string
    delete:
      tags:
        - "User"
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Succesfully deletes one user"
          schema:
            type: object
            properties:
              acknowledged:
                type: boolean
              deletedCount:
                type: integer
        "400":
          description: Returns error
          schema:
            type: object
            properties:
              message:
                type: string
  /auth/login:
    get:
      tags:
        - "Auth"
      parameters:
        - in: body
          name: userCredentials
          description: User's credentials.
          schema:
            type: object
            properties:
              username:
                type: string
              password:
                type: string
      responses:
        "200":
          description: "Succesfully log in User"
          schema:
            type: object
            properties:
              _id:
                type: string
              username:
                type: string
              password:
                type: string
              first_name:
                type: string
              last_name:
                type: string
              profile_photo_path:
                type: string
              bio:
                type: string
              phone_number:
                type: string
              __v:
                type: integer
        "400":
          description: Returns error
          schema:
            type: object
            properties:
              message:
                type: string
        "404":
          description: Returns error
          schema:
            type: object
            properties:
              message:
                type: string
  /auth/register:
    post:
      tags:
        - "Auth"
      parameters:
        - in: body
          name: userInfo
          description: User's attributes.
          schema:
            type: object
            properties:
              username:
                type: string
              password:
                type: string
              first_name:
                type: string
              last_name:
                type: string
      responses:
        "201":
          description: "Succesfully register user"
          schema:
            type: object
            properties:
              username:
                type: string
              password:
                type: string
              first_name:
                type: string
              last_name:
                type: string
              profile_photo_path:
                type: string
              bio:
                type: string
              phone_number:
                type: string
              __v:
                type: integer
        "406":
          description: Returns error
          schema:
            type: object
            properties:
              message:
                type: string
        "500":
          description: Returns error
          schema:
            type: object
            properties:
              message:
                type: string
  /categories:
    get:
      tags:
        - "Categories"
      responses:
        "200":
          description: "Succesfully returns all categories"
          schema:
            type: object
            properties:
              _id:
                type: string
              name:
                type: string
              __v:
                type: integer
        "404":
          description: Returns error
          schema:
            type: object
            properties:
              message:
                type: string
  /categories/{categoryId}:
    get:
      tags:
        - "Categories"
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Succesfully returns one category"
          schema:
            type: object
            properties:
              _id:
                type: string
              name:
                type: string
              __v:
                type: integer
        "400":
          description: Returns error
          schema:
            type: object
            properties:
              message:
                type: string
  /activities:
    post:
      tags:
        - "Activity"
      parameters:
        - in: body
          name: activity
          description: The activity's attributes.
          schema:
            type: object
            properties:
              category_id:
                type: string
              date:
                type: Date
              place:
                type: string
              share_phone:
                type: Boolean
              as_group:
                type: Boolean
              user_id:
                type: string
              group_id:
                type: string
              description:
                type: string
      responses:
        "200":
          description: "Succesfully creates activity"
          schema:
            type: object
            properties:
              _id:
                type: string
              category_id:
                type: string
              date:
                type: Date
              place:
                type: string
              share_phone:
                type: Boolean
              as_group:
                type: Boolean
              user_id:
                type: string
              group_id:
                type: string
              description:
                type: string
              __v:
                type: string
        "404":
          description: Returns error
          schema:
            type: object
            properties:
              message:
                type: string             
  /activities/{activityId}:
    get:
      tags:
        - "Activity"
      parameters:
        - in: path
          name: activityId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: "Succesfully returns one activity with provided id"
          schema:
            type: object
            properties:
              _id:
                type: string
              category_id:
                type: string
              date:
                type: Date
              place:
                type: string
              share_phone:
                type: Boolean
              as_group:
                type: Boolean
              user_id:
                type: string
              group_id:
                type: string
              description:
                type: string
              __v:
                type: string
        "404":
          description: Returns error
          schema:
            type: object
            properties:
              message:
                type: string
    delete:
      tags:
        - "Activity"
      parameters:
        - in: path
          name: activityId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: "Succesfully deletes one activity"
          schema:
            type: object
            properties:
              acknowledged:
                type: boolean
              deletedCount:
                type: integer
        "404":
          description: Returns error
          schema:
            type: object
            properties:
              message:
                type: string
  /activities/user/{userId}:
    get:
      tags:
        - "Activity"
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: "Succesfully returns activities of provided user"
          schema:
            type: object
            properties:
              _id:
                type: string
              category_id:
                type: string
              date:
                type: Date
              place:
                type: string
              share_phone:
                type: Boolean
              as_group:
                type: Boolean
              user_id:
                type: string
              group_id:
                type: string
              description:
                type: string
              __v:
                type: string
        "404":
          description: Returns error
          schema:
            type: object
            properties:
              message:
                type: string
  /groups:
    post:
      tags:
        - "Group"
      parameters:
        - in: body
          name: activity
          description: The group's attributes.
          schema:
            type: object
            properties:
              name:
                type: string
              profile_photo_path:
                type: Date
              bio:
                type: string
      responses:
        "200":
          description: "Succesfully creates group"
          schema:
            type: object
            properties:
              _id:
                type: string
              name:
                type: string
              profile_photo_path:
                type: Date
              place:
                type: string
              share_phone:
                type: Boolean
              as_group:
                type: Boolean
              user_id:
                type: string
              group_id:
                type: string
              description:
                type: string
              __v:
                type: string
        "404":
          description: Returns error
          schema:
            type: object
            properties:
              message:
                type: string 
  /groups/{groupId}:
    get:
      tags:
        - "Group"
      parameters:
        - in: path
          name: groupId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: "Succesfully returns one group with provided id"
          schema:
            type: object
            properties:
              _id:
                type: string
              name:
                type: string
              profile_photo_path:
                type: Date
              bio:
                type: string
              __v:
                type: string
        "404":
          description: Returns error
          schema:
            type: object
            properties:
              message:
                type: string
    delete:
      tags:
        - "Group"
      parameters:
        - in: path
          name: groupId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: "Succesfully deletes group"
          schema:
            type: object
            properties:
              acknowledged:
                type: boolean
              deletedCount:
                type: integer
        "404":
          description: Returns error
          schema:
            type: object
            properties:
              message:
                type: string
  /invites:
    post:
      tags:
        - "Invite"
      parameters:
        - in: body
          name: inviteInfo
          description: The invite's attributes.
          schema:
            type: object
            properties:
              group_id:
                type: string
              user_id:
                type: string
      responses:
        "200":
          description: "Succesfully creates invite"
          schema:
            type: object
            properties:
              group_id:
                type: string
              user_id:
                type: string
              _id:
                type: string
              __v:
                type: integer
        "404":
          description: Returns error
          schema:
            type: object
            properties:
              message:
                type: string
  /invites/{inviteId}:
    get:
      tags:
        - "Invite"
      parameters:
        - in: path
          name: inviteId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Succesfully returns one invite with provided id"
          schema:
            type: object
            properties:
              group_id:
                type: string
              user_id:
                type: string
              _id:
                type: string
              __v:
                type: integer
        "404":
          description: Returns error
          schema:
            type: object
            properties:
              message:
                type: string
    delete:
      tags:
        - "Invite"
      parameters:
        - in: path
          name: inviteId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Succesfully deletes invite"
          schema:
            type: object
            properties:
              acknowledged:
                type: boolean
              deletedCount:
                type: integer
        "404":
          description: Returns error
          schema:
            type: object
            properties:
              message:
                type: string
  /invites/user/{userId}:
    get:
      tags:
        - "Invite"
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Succesfully returns all user's invites"
          schema:
            type: object
            properties:
              group_id:
                type: string
              user_id:
                type: string
              _id:
                type: string
              __v:
                type: integer
        "404":
          description: Returns error
          schema:
            type: object
            properties:
              message:
                type: string
  /usersActivities:
    post:
      tags:
        - "UsersActivities"
      parameters:
        - in: body
          name: recordInfo
          description: Id of user and activity logging into.
          schema:
            type: object
            properties:
              user_id:
                type: string
              activity_id:
                type: string
      responses:
        "200":
          description: "Succesfully creates usersActivity record"
          schema:
            type: object
            properties:
              user_id:
                type: string
              activity_id:
                type: string
              _id:
                type: string
              __v:
                type: integer
        "404":
          description: Returns error
          schema:
            type: object
            properties:
              message:
                type: string
  /usersActivities/{Id}:
    delete:
      tags:
        - "UsersActivities"
      parameters:
        - in: path
          name: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Succesfully deletes usersActivities record"
          schema:
            type: object
            properties:
              acknowledged:
                type: boolean
              deletedCount:
                type: integer
        "400":
          description: Returns error
          schema:
            type: object
            properties:
              message:
                type: string
  /usersGroups:
    post:
      tags:
        - "UsersGroups"
      parameters:
        - in: body
          name: recordInfo
          description: Id of user and group logging into.
          schema:
            type: object
            properties:
              user_id:
                type: string
              group_id:
                type: string
      responses:
        "200":
          description: "Succesfully creates usersGroups record"
          schema:
            type: object
            properties:
              user_id:
                type: string
              group_id:
                type: string
              _id:
                type: string
              __v:
                type: integer
        "404":
          description: Returns error
          schema:
            type: object
            properties:
              message:
                type: string
  /usersGroups/{Id}:
    delete:
      tags:
        - "UsersGroups"
      parameters:
        - in: path
          name: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Succesfully deletes usersGroups record"
          schema:
            type: object
            properties:
              acknowledged:
                type: boolean
              deletedCount:
                type: integer
        "400":
          description: Returns error
          schema:
            type: object
            properties:
              message:
                type: string
  /users/activity/{activityId}:
    get:
      tags:
        - "User"
      parameters:
        - in: path
          name: activityId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Succesfully returns all users of the provided activity"
          schema:
            type: object
            properties:
              _id:
                type: string
              username:
                type: string
              password:
                type: string
              first_name:
                type: string
              last_name:
                type: string
              profile_photo_path:
                type: string
              bio:
                type: string
              phone_number:
                type: string
              __v:
                type: integer
        "404":
          description: Returns error
          schema:
            type: object
            properties:
              message:
                type: string
  /users/group/{groupId}:
    get:
      tags:
        - "User"
      parameters:
        - in: path
          name: groupId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Succesfully returns all users of the provided group"
          schema:
            type: object
            properties:
              _id:
                type: string
              username:
                type: string
              password:
                type: string
              first_name:
                type: string
              last_name:
                type: string
              profile_photo_path:
                type: string
              bio:
                type: string
              phone_number:
                type: string
              __v:
                type: integer
        "404":
          description: Returns error
          schema:
            type: object
            properties:
              message:
                type: string
  /groups/user/{userId}:
    get:
      tags:
        - "Group"
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: "Succesfully returns all groups of the provided user"
          schema:
            type: object
            properties:
              _id:
                type: string
              name:
                type: string
              profile_photo_path:
                type: Date
              bio:
                type: string
              __v:
                type: string
        "404":
          description: Returns error
          schema:
            type: object
            properties:
              message:
                type: string
  /activities/group/{groupId}:
    get:
      tags:
        - "Activity"
      parameters:
        - in: path
          name: groupId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: "Succesfully returns activities of provided group"
          schema:
            type: object
            properties:
              _id:
                type: string
              category_id:
                type: string
              date:
                type: Date
              place:
                type: string
              share_phone:
                type: Boolean
              as_group:
                type: Boolean
              user_id:
                type: string
              group_id:
                type: string
              description:
                type: string
              __v:
                type: string
        "404":
          description: Returns error
          schema:
            type: object
            properties:
              message:
                type: string